flask_app/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
soupsieve==1.7.3
urllib3==1.24.1
Click==7.0
Flask==1.0.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Werkzeug==0.14.1
WTForms==2.2.1
gunicorn==19.9.0
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.4.0

